src/daily/L451/SolutionTest.java
package daily.L451;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

class SolutionTest {
    @Test
    public void case1() {
        Solution solution = new Solution();
        Assertions.assertEquals(solution.frequencySort("tree"), "eert");
    }

    @Test
    public void case2() {
        Solution solution = new Solution();
        Assertions.assertEquals(solution.frequencySort("cccaaa"), "aaaccc");
    }

    @Test
    public void case3() {
        Solution solution = new Solution();
        Assertions.assertEquals(solution.frequencySort("Aabb"), "bbAa");
    }
}